Split analytics into multiple components
